serv/requirement.txt
pyserial
fastapi
uvicorn
